+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 * You can now use `Date#startOf(unit)` to set dacesecond & decaminute
 * `Function.series` will now pass the value from the previous task to the next
 * Add `LazyPledge` class, which will only call on the executor once `then` has been called
+* Forward Pledge progress reports to other pledges
 
 ## 0.6.2 (2018-11-09)