+++ CHANGELOG.md
@@ -8,6 +8,7 @@
 * Add `Blast.isWebview` to detect when it's running in a webview
 * Use `requestIdleCallback` as a `setImmediate` implementation in the browser
 * Make `Informer#emit()` exit early when no listeners are found
+* Rewrite setImmediate polyfill
 
 ## 0.6.2 (2018-11-09)