+++ CHANGELOG.md
@@ -1,4 +1,4 @@
-## 0.6.3 (WIP)
+## 0.6.3 (2018-12-06)
 
 * Add `RURL#usedBaseProperty(name)` to find out if a property came from the base location
 * You can now use `Date#startOf(unit)` to set dacesecond & decaminute